Declare index as size_t when looping on primes

+++ src/main/cpp/app2/sieve.cpp
@@ -81,7 +81,7 @@ void Sieve::compute_primes() {
 void Sieve::print_primes() const {
   std::cout << "Prime numbers in range 0-" << this->range << " inclusive:" << std::endl;
   int half = this->primes.size() / 2;
-  for (int n = 0; n < this->primes.size(); n++) {
+  for (size_t n = 0; n < this->primes.size(); n++) {
     int p = this->primes.at(n);
     if (half > MAX_HALF) {
       if (n == MAX_HALF) {
